requirements.txt
fastmcp>=0.1.0
python-dotenv==1.0.0
yandex_tracker_client>=2.6.0 